cart detail page remainings

+++ src/Shared/BlazorBoilerplate.Infrastructure.Storage/DataModels/BBShop/ProductStats.cs
@@ -8,8 +8,8 @@ namespace BlazorBoilerplate.Infrastructure.Storage.DataModels
     
     public partial class ProductStats : IAuditable, ISoftDelete
     {
-        public int Id { get; set; }
-        public int Visits { get; set; }
+        public long Id { get; set; }
+        public long Visits { get; set; }
         public DateTime? LastVisit { get; set; }
 
     }
